# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apricode documentation",
        "description": "V1 Api docs",
        "contact": {
            "email": "ialeroy@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Список жанров",
                "description": "Список жанров",
                "operationId": "66e9590ab65c5c762d3cc4d85bbebc21",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "count",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Genre"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Добавить жанр",
                "description": "Добавить жанр",
                "operationId": "66af595f201b38027951516bac71ed6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/genres/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Страница жанра",
                "description": "Страница жанра",
                "operationId": "7cb562b914ccd8e03bcd0ab72affbf8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Удаление жанра",
                "description": "Удаление жанра",
                "operationId": "08d4bad77bb64e72fc79402fac1889e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Genres"
                ],
                "summary": "Обновление жанра",
                "description": "Обновление жанра",
                "operationId": "f2e4ff2e4875d7776bb24dbe0cce9895",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Genre": {
                "title": "Genre",
                "description": "Genre Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Имя",
                        "type": "string",
                        "example": "Rachelle Homenick"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Создана в",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Genre"
                }
            },
            "Pagination": {
                "title": "Pagination",
                "description": "Pagination",
                "properties": {
                    "current_page": {
                        "title": "current_page",
                        "description": "current_page",
                        "type": "integer"
                    },
                    "from": {
                        "title": "from",
                        "description": "from",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "last_page",
                        "description": "last_page",
                        "type": "integer"
                    },
                    "path": {
                        "title": "path",
                        "description": "path",
                        "type": "string"
                    },
                    "per_page": {
                        "title": "per_page",
                        "description": "per_page",
                        "type": "integer"
                    },
                    "to": {
                        "title": "to",
                        "description": "to",
                        "type": "integer"
                    },
                    "total": {
                        "title": "total",
                        "description": "total",
                        "type": "integer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Pagination"
                }
            },
...
            "GenreRequest": {
                "title": "Genre request",
                "description": "Genre request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "example": "A nice Genre"
                    }
                },
                "type": "object"
            }
        }
    },
...
}